Update 11/05/2024
NEW: Inventory system done!!!!
Needs more item in the future, currently we only have item to increase HP and SP for the character

+++ Assets/Scripts/UI/Player HUD/SPBar.cs
@@ -27,7 +27,7 @@ public class SPBar : MonoBehaviour
         }
     }
 
-    public void SetSP(int currentSP, int maxSP)
+    public void SetSP(float currentSP, int maxSP)
     {
         // Set the SP value and the max SP value of the SP bar
         spBar.value = currentSP;
